+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changes by Version
 - Fixed a bug where the 'not found' handler would incorrectly return
   serialization mismatch errors..
 - Fixed a bug which prevented VCR support from working with the sync client.
+- Fixed a bug in VCR that prevented it from recording requests made by the sync
+  client, and requests made with ``hostport=None``.
 - Made ``client_for`` compatible with ``tchannel.TChannel``.
 - Brought back ``tchannel.sync.client_for`` for backwards compatibility.